+++ CdnProAssetBundle.php
@@ -1,0 +1,22 @@
+<?php
+/**
+ * CdnProAssetBundle.php
+ * @author Revin Roman
+ * @link https://rmrevin.ru
+ */
+
+namespace rmrevin\yii\fontawesome;
+
+/**
+ * Class CdnProAssetBundle
+ * @package rmrevin\yii\fontawesome
+ */
+class CdnProAssetBundle extends \yii\web\AssetBundle
+{
+    /**
+     * @inherit
+     */
+    public $css = [
+        'https://pro.fontawesome.com/releases/v5.3.1/css/all.css',
+    ];
+}